- fix bug, for static cursor clear sqlind before fetch branch v1-2-beta

+++ IscDbc/TypesResultSet.h
@@ -45,7 +45,7 @@ public:
 	int			recordNumber;
 	int			dataTypes;
 	Sqlda		outputSqlda;
-	short		*indicators;
+	long		*indicators;
 };
 
 }; // end namespace IscDbcLibrary
